refactor: 패키지 분리, 클래스 이름 변경
