+++ src/main/java/racingcar/view/InputView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,17 +10,17 @@ public class InputView {
 
     private final InputParser inputParser = new InputParser();
     private final Scanner scanner = new Scanner(System.in);
-    private final OutputView outputView;
+    private final ResultView resultView;
 
-    public InputView(OutputView outputView) {
-        this.outputView = outputView;
+    public InputView(ResultView resultView) {
+        this.resultView = resultView;
     }
 
     private final int MINIMUM_CARS = 2;
 
 
     public List<Car> getCars() {
-        outputView.printMessage("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
+        resultView.printMessage("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         String carNamesInput = scanner.nextLine();
         String[] carNames = inputParser.splitByComma(carNamesInput);
         validateCarNames(carNames);
@@ -30,7 +32,7 @@ public class InputView {
     }
 
     public int getRounds() {
-        outputView.printMessage("시도할 회수는 몇회인가요?");
+        resultView.printMessage("시도할 회수는 몇회인가요?");
         String roundInput = scanner.nextLine();
         return inputParser.parseToInt(roundInput);
     }
