+++ src/main/java/racingcar/view/ResultView.java
@@ -1,8 +1,10 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Cars;
 
 import java.util.List;
 
-public class OutputView {
+public class ResultView {
 
     public void printMessage(String message) {
         System.out.println(message);
