+++ src/test/java/racingcar/CarsTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 import java.util.Arrays;
 import java.util.List;
